[Playground] [Backend] removed cloud storage
